graph-service/requirements.txt
motor~=2.5.1
PyJWT~=2.4.0
bcrypt
uvicorn~=0.17.6
pymongo~=3.12.3
fastapi~=0.78.0
pydantic[email,dotenv]~=1.9.0
passlib~=1.7.4
kafka-python~=2.0.2
beanie~=1.10.8
brokerstream==0.0.8
sse-starlette==1.1.6
asyncio==3.4.3
openai
translators
googletrans
google